Initial import of compressed repo history
This initial command line client is able to help diagnose routing easily
for people confused by ping and traceroute utilities along with NAT.

The main purpose of it is to monitor your external IP and update Parvati
when it changes. It also looks for your host and auto-announces you.

Code is still early alpha.